+++ crates/rpc-engine-types/Cargo.toml
@@ -1,0 +1,44 @@
+[package]
+name = "alloy-rpc-engine-types"
+description = "Alloy RPC engine types"
+
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+authors.workspace = true
+repository.workspace = true
+exclude.workspace = true
+
+[dependencies]
+# ethereum
+alloy-rlp = { workspace = true, features = ["arrayvec", "derive"] }
+alloy-primitives = { workspace = true, features = ["rlp", "serde"] }
+alloy-rpc-types.workspace = true
+
+itertools.workspace = true
+serde = { workspace = true, features = ["derive"] }
+serde_json.workspace = true
+thiserror.workspace = true
+
+# arbitrary
+arbitrary = { version = "1.3", features = ["derive"], optional = true }
+proptest = { version = "1.4", optional = true }
+proptest-derive = { version = "0.4", optional = true }
+
+# jsonrpsee
+jsonrpsee-types = { version = "0.20", optional = true }
+
+[features]
+arbitrary = ["dep:arbitrary", "dep:proptest-derive", "dep:proptest", "alloy-primitives/arbitrary"]
+jsonrpsee-types = ["dep:jsonrpsee-types"]
+
+[dev-dependencies]
+alloy-primitives = { workspace = true, features = ["rand", "rlp", "serde", "arbitrary"] }
+
+arbitrary = { workspace = true, features = ["derive"] }
+proptest.workspace = true
+proptest-derive.workspace = true
+rand.workspace = true
+similar-asserts.workspace = true
