+++ crates/rpc-engine-types/src/forkchoice.rs
@@ -1,5 +1,5 @@
 use super::{PayloadStatus, PayloadStatusEnum};
-use crate::engine::PayloadId;
+use crate::PayloadId;
 use alloy_primitives::B256;
 use serde::{Deserialize, Serialize};
